Correct for new protected_attributes and constrain to rails ~>4.0
The previous assumption that this gem could cover all of Rails versions
>= 3.0, < 5.0 forces us to work around the differences between rails 3.x
mass assignment protection and rails 4.x usage of protected_attributes.

By restricting to ~>4.0 for this version of the gem, we can be sure that
if assignment protection is being used, it is via protected_attributes,
and therefore use the three-argument API for sanitize. dm-rails 1.2
will still handle the previous rails 3 two-argument sanitize API.